Exiv2 v1.0.0 compatibility changes

+++ src/swig_0.27.0/image_wrap.cxx
@@ -6021,7 +6021,7 @@ SWIGINTERN PyObject *_wrap_ExifDataIterator_copy(PyObject *self, PyObject *args)
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
@@ -7348,7 +7348,7 @@ SWIGINTERN PyObject *_wrap_IptcDataIterator_copy(PyObject *self, PyObject *args)
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
@@ -8870,7 +8870,7 @@ SWIGINTERN PyObject *_wrap_XmpDataIterator_copy(PyObject *self, PyObject *args) 
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
