+++ src/swig_0.27.4/exif_wrap.cxx
@@ -5611,7 +5611,7 @@ SWIGINTERN PyObject *_wrap_ExifDataIterator_copy(PyObject *self, PyObject *args)
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
@@ -7117,7 +7117,7 @@ SWIGINTERN PyObject *_wrap_Exifdatum_copy(PyObject *self, PyObject *args) {
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
