+++ src/swig_0.27.4/xmp_wrap.cxx
@@ -5122,7 +5122,7 @@ SWIGINTERN PyObject *_wrap_XmpDataIterator_copy(PyObject *self, PyObject *args) 
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
@@ -6407,7 +6407,7 @@ SWIGINTERN PyObject *_wrap_Xmpdatum_copy(PyObject *self, PyObject *args) {
   void *argp1 = 0 ;
   int res1 = 0 ;
   Py_buffer view2 ;
-  int _global_len ;
+  long _global_len ;
   int val3 ;
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
